// src/pages/orderAnalytics/types/order.type.ts
export type OrderItem = {
    order_item_id: string;
    order_id: string;
    fulfilment_source: string;
    fulfilment_type: string;
    order_date: string;
    order_approval_date: string;
    order_item_status: string;
    sku: string;
    fsn: string;
    product_title: string;
    quantity: string;
    serial_no_imei: string;
    delivery_logistics_partner: string;
    pickup_logistics_partner: string;
    delivery_tracking_id: string;
    forward_logistics_form: string;
    forward_logistics_form_no: string;
    order_cancellation_date: string;
    cancellation_reason: string;
    cancellation_sub_reason: string;
    order_return_approval_date: string;
    return_id: string;
    return_reason: string;
    return_sub_reason: string;
    procurement_dispatch_sla: string;
    dispatch_after_date: string;
    dispatch_by_date: string;
    order_ready_for_dispatch_on_date: string;
    dispatched_date: string;
    dispatch_sla_breached: string;
    seller_pickup_reattempts: string;
    delivery_sla: string;
    deliver_by_date: string;
    order_delivery_date: string;
    delivery_sla_breached: string;
    order_service_completion_date: string;
    service_by_date: string;
    service_completion_sla: string;
    service_sla_breached: string;
  };
  